update CuptiCounterAnalysis to use the new trace call graph implementation
